Continued on BitcoinWallet test implementation and added datadir spoofing to BitcoinWallet. Also fixed tabs in lib.processing.

+++ tests/integration/plugins_currencies_bitcoin.py
@@ -5,20 +5,24 @@
 # Python
 import unittest
 import os
+import sys
 import shutil
+from subprocess import Popen, PIPE
 from pathlib import Path
 
 # Local
 from tests.lib.mocking import DummyProcess
 
 # What's to be tested.
-from plugins.currencies.bitcoin import BitcoinWallet
+from plugins.currencies.bitcoin import\
+	BitcoinWallet as Wallet,\
+	BitcoinConfig as Config
 
 #=======================================================================================
 # Configuration
 #=======================================================================================
 
-DATA_DIR = str(Path(Path(__file__).absolute().parent, "tests", "data", "plugins_currency_bitcoin"))
+TESTDATA_DIR = str(Path(Path(sys.argv[0]).resolve().parent, "tests", "data", "plugins_currency_bitcoin"))
 
 #=======================================================================================
 # Tests
@@ -38,11 +42,18 @@ class BitcoinMockProcess(DummyProcess):
 			if homeDirSourcePath is None:
 				self.homeDirPath.mkdir(exist_ok=True)
 			else:
-				shutil.copytree(str(self.homeDirSourcePath), str(self.homeDirPath))
+				if self.homeDirSourcePath.exists():
+					shutil.copytree(str(self.homeDirSourcePath), str(self.homeDirPath))
 	
+	@property
 	def homeDirPath(self):
 		"""Path object for the home directory."""
-		return Path(tempDir.name, "home")
+		return Path(self.tempDir.name, "home")
+	
+	@property
+	def basePath(self):
+		"""Return the path of the directory the binaries are expected to reside in."""
+		return Path(self.execPath).parent
 	
 	@property
 	def process(self):
@@ -71,29 +82,44 @@ class BitcoinTestCase(unittest.TestCase):
 	
 	prefix = "blockchaintools_test"
 	
-	@property
-	def mockDaemonSourcePath(self):
-		return Path(DATA_DIR, "bitcoind.c")
-	
-	@property
-	def mockCliSourcePath(self):
-		return Path(DATA_DIR, "bitcoin-cli.c")
-	
-	@property
-	def homeDirPath(self):
-		return Path(DATA_DIR, "default-home")
-	
 	def __init__(self, methodName="runTest"):
 		super().__init__(methodName)
 		self.testDaemonProcess = BitcoinDaemonMockProcess(sourcePath=str(self.mockDaemonSourcePath),\
-			homeDirSourcePath=str(self.homeDirPath))
+			homeDirSourcePath=str(self.homeDirSourcePath))
 		self.testCliProcess = BitcoinCliMockProcess(sourcePath=str(self.mockCliSourcePath),\
-			homeDirSourcePath=str(self.homeDirPath))
-
-#class BitcoinDaemonTestCase(BitcoinTestCase):
+			homeDirSourcePath=str(self.homeDirSourcePath))
+	
+	@property
+	def mockDaemonSourcePath(self):
+		return Path(TESTDATA_DIR, "bitcoind.c")
+	
+	@property
+	def mockCliSourcePath(self):
+		return Path(TESTDATA_DIR, "bitcoin-cli.c")
+	
+	@property
+	def homeDirSourcePath(self):
+		return Path(TESTDATA_DIR, "default-home")
+	
+	def setUp(self):
+		self.testDaemonProcess.start()
+		
+	def tearDown(self):
+		self.testDaemonProcess.stop()
+	
+class BitcoinDaemonTestCase(BitcoinTestCase):
+	
+	def newWalletInstance(self):
+		return Wallet(Config(\
+			basePaths=[str(self.testDaemonProcess.basePath)],\
+			dataDirBaseDirPath=self.testDaemonProcess.homeDirPath))
+	
 	#def test_start(self):
-		#pass#TODO
-		#BitcoinWallet(basePaths=Path(self.testDaemonProcess.execPath).parent)
+	#	wallet = self.newWalletInstance()
+	
+	def test_getDaemon(self):
+		wallet = self.newWalletInstance()
+		wallet.getDaemon()
 
 if __name__ == "__main__":
 	unittest.main()
